requirements.txt
aniso8601==9.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
Deprecated==1.2.12
distlib==0.3.2
filelock==3.0.12
firebase-admin==5.0.1
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Pydantic==0.8.0
flask-pydantic-spec==0.1.4
flask-restplus==0.13.0
flask-swagger-ui==3.36.0
future==0.18.2
gcloud==0.18.3
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
inflection==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jwcrypto==1.0
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
mistune==0.8.4
msgpack==1.0.2
nested-lookup==0.2.21
numpy==1.21.1
oauth2client==4.1.3
packaging==21.0
pandas==1.3.1
pipenv==2021.5.29
platformdirs==2.1.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
Pyrebase4==4.5.0
pyrsistent==0.18.0
pythainlp==2.3.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
spectree==0.6.1
threadpoolctl==2.2.0
tinydb==4.5.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.6.0
virtualenv-clone==0.5.5
Werkzeug==2.0.1
wrapt==1.12.1
XlsxWriter==3.0.1
